+++ src/colorer/strings/legacy/Encodings.cpp
@@ -180,7 +180,7 @@ uUnicodeString Encodings::fromUTF8(char* data, int32_t len)
 uUnicodeString Encodings::fromUTF8(unsigned char* data)
 {
   const auto c = reinterpret_cast<char*>(data);
-  return fromUTF8(c,strlen(c));
+  return fromUTF8(c,(int32_t)strlen(c));
 }
